mononoke/hgcli: log hgcli server data to Scuba
Summary:
This will give us the host that hgcli is running on, which is, like, convenient
to know what hg host proxied a request. For comparison, currently, we have to
go through ssh prod logs using the client IP and port, and hope there aren't
too many matches, which is really not a reasonable way to debug things.

Reviewed By: ahornby

Differential Revision: D23994304

fbshipit-source-id: fa5b29aa50e278f0f1b3b60be42f634a1c5c45c1