get rid of synced_commit_mapping::get_one in admin
Summary:
`get_one` is a deprecated method, because it uses incorrect logic to resolve ambiguities of multi-mapped commits: if just selects the very first of the potentially many mappings.

Correct resolution is to either handle the ambiguity at the caller site, or rely on provided resolution logic in commit_sync_outcome.rs.

Therefore, I am removing the uses of this method in this and a few surrounding commits.

In this case, the simplest thing is to replace it with `.get` and deal with multi-mappings on the client side:
- for `crossrepo map` subcommand we just print all mappings
- for `update_large_repo_bookmarks` we just fail on multi-mapping, as it seems dangerous to proceed without human intervention

Reviewed By: farnz

Differential Revision: D24030033

fbshipit-source-id: c84613579fbf8a5f6bac3c06da0cd4e0ad6c3fb0