gotham_ext: add FuseOnErr stream combinator
Summary:
Add a stream combinator that causes a `TryStream` to end upon encountering the error. The resulting stream remains a `TryStream`, but will always terminate after yielding an error item.

Later in this stack, this will be used in the LFS server to cause mid-stream errors to always terminate the response.

Reviewed By: krallin

Differential Revision: D24301254

fbshipit-source-id: e71ddc16e79caeb195e23be7d6a93ee63a96a713