gotham_ext: use pin_project in SignalStream
Summary: This code predates the use of `pin_project`, which simplifies the implementation of pinned types using proc macros. We now use `pin_project` in most other places in this crate where we define custom `Future`s and `Stream`s, so let's use it in `SignalStream` as well. (This makes the code easier to modify and removes the need for `unsafe` code.)

Reviewed By: farnz

Differential Revision: D24285792

fbshipit-source-id: da018b9ed62c33d9d06ecf6caa4748969a057e66