mononoke/test-redaction.t: give jq a program to avoid surprises
Summary:
We don't give `jq` a program here, and it looks like depending on the exact
setup, this might result in it printing the help instead of running, as
reported by Jun here:
https://fb.workplace.com/groups/scm.mononoke/permalink/1280281002334472/

I think this is the behaviro inquestion:
https://github.com/stedolan/jq/issues/1110. I'm guessing something might have
changed in jq 1.6 because I am on that version and don't get that behavior but
Jun's output uses jq 1.5.

Let's give jq a program to fix this.

Reviewed By: StanislavGlebik

Differential Revision: D24329472

fbshipit-source-id: f4ba4246770dd79630433a3c46c8bbd38aa4c842