mononoke_types-unittest: don't depend on fixtures to get an empty blobstore
Summary:
Those tests need an empty blobstore, but to do this, they `fixtures::linear`,
which a) instantiates a whole blobrepo, and b) fill it with commits we never
use.

In particular, a) is kinda annoying, because that creates a dependency on
blobrepo and therefore on most of Mononoke. This makes the tests much slower to
build than they could be.

This diff updates the test to just get their memblob directly instead of via a
fixture, thus bringing down the number of Mononoke crates we pull in from 120
to 21.

Reviewed By: StanislavGlebik

Differential Revision: D24305564

fbshipit-source-id: 91c9423f6284f985461744d9c61ce207024a65c8